requirements.txt
loguru
python-dotenv
selenium